Add a Test.DejaFu.Settings module
